Added all the results data to the report, still to comment it. Fixed scripts to produce CSV files to be converted with csv2latex. Heuristic init function working.

+++ bin/launchTSPTW-VND.sh
@@ -4,8 +4,11 @@
 # $3 = Seeds output file
 
 createReportFile() {
-	touch $1
-	echo "Instance	Infeasible	mean(PRDP)	mean(CpuTime)" > $1
+	if [ ! -f $1 ]
+	then
+		touch $1
+		echo "Instance	Infeasible	mean(PRDP)	mean(CpuTime)" > $1
+	fi
 }
 
 generateRandomSeeds() {
